+++ packages/client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # fuzd-client
 
+## 0.9.6
+
+### Patch Changes
+
+- crititialDelta
+- Updated dependencies
+  - fuzd-scheduler@0.9.5
+  - fuzd-executor@0.9.5
+  - fuzd-common@0.9.4
+  - fuzd-chain-protocol@0.9.5
+
 ## 0.9.5
 
 ### Patch Changes