+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # fuzd-executor
 
+## 0.9.4
+
+### Patch Changes
+
+- crititialDelta
+
 ## 0.9.3
 
 ### Patch Changes