+++ packages/tlock-decrypter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # fuzd-tlock-decrypter
 
+## 0.9.5
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.9.5
+  - fuzd-common@0.9.4
+
 ## 0.9.4
 
 ### Patch Changes